requirements.txt
asgiref
Django
djangorestframework
pillow
pip
psycopg2-binary
pytz
setuptools
sqlparse
tzdata
faker